Refactor external auth validators
