+++ api/Avancira.Infrastructure/Auth/IExternalTokenValidator.cs
@@ -1,0 +1,9 @@
+using Avancira.Application.Auth;
+
+namespace Avancira.Infrastructure.Auth;
+
+public interface IExternalTokenValidator
+{
+    string Provider { get; }
+    Task<ExternalAuthResult> ValidateAsync(string token);
+}
